test(core): split mocks_test.go into focused test files
